[EZ][TP] Refactor FSDP 2D integration extension code so that it can re-used
[ghstack-poisoned]
